Create matrix_diagona_sum.py
Initialize an empty list of  primary diagonal and secondary diagonal. In a list of list for primary diagonal access from first element to last element. For secondary diagonal acces elements from last. Then sum primary diagonal and secondary diagonal. 

Note: If the matrix length is odd you have to find the mid of mid element and subtract it to the total sum.